+++ xenon-common/src/main/java/com/vmware/xenon/common/SynchronizationTaskService.java
@@ -37,6 +37,8 @@ public class SynchronizationTaskService
     public static final String FACTORY_LINK = ServiceUriPaths.SYNCHRONIZATION_TASKS;
     public static final String PROPERTY_NAME_SYNCHRONIZATION_LOGGING = Utils.PROPERTY_NAME_PREFIX
             + "SynchronizationTaskService.isDetailedLoggingEnabled";
+    public static final String PROPERTY_NAME_ENABLE_ODL_SYNCHRONIZATION = Utils.PROPERTY_NAME_PREFIX
+            + "enableOdlSynchronization";
 
     public static final String STAT_NAME_CHILD_SYNCH_RETRY_COUNT = "childSynchRetryCount";
     public static final String STAT_NAME_SYNCH_RETRY_COUNT = "synchRetryCount";
@@ -328,8 +330,12 @@ public class SynchronizationTaskService
             return null;
         }
         if (currentTask.childOptions.contains(ServiceOption.ON_DEMAND_LOAD)) {
-            put.fail(new IllegalArgumentException("ON_DEMAND_LOAD services must synchronize on-demand."));
-            return null;
+            boolean odlSync =
+                    Boolean.valueOf(System.getProperty(SynchronizationTaskService.PROPERTY_NAME_ENABLE_ODL_SYNCHRONIZATION));
+            if (!odlSync) {
+                put.fail(new IllegalArgumentException("ON_DEMAND_LOAD services must synchronize on-demand."));
+                return null;
+            }
         }
         if (currentTask.membershipUpdateTimeMicros != null &&
                 currentTask.membershipUpdateTimeMicros > putTask.membershipUpdateTimeMicros) {