+++ xenon-common/src/test/java/com/vmware/xenon/common/TestSynchronizationTaskService.java
@@ -18,6 +18,7 @@ import static java.util.stream.Collectors.toList;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 import java.net.URI;
@@ -111,6 +112,8 @@ public class TestSynchronizationTaskService extends BasicTestCase {
     public static void tearDownClass() throws Exception {
         System.setProperty(
                 SynchronizationTaskService.PROPERTY_NAME_SYNCHRONIZATION_LOGGING, "false");
+        System.setProperty(
+                SynchronizationTaskService.PROPERTY_NAME_ENABLE_ODL_SYNCHRONIZATION, "false");
     }
 
     @Override
@@ -122,6 +125,7 @@ public class TestSynchronizationTaskService extends BasicTestCase {
                 InMemoryLuceneDocumentIndexService.SELF_LINK);
 
         this.host.startFactory(InMemoryExampleService.class, InMemoryExampleService::createFactory);
+        this.host.startFactory(ExampleODLService.class, ExampleODLService::createFactory);
     }
 
     @Before
@@ -129,8 +133,12 @@ public class TestSynchronizationTaskService extends BasicTestCase {
         CommandLineArgumentParser.parseFromProperties(this);
         URI exampleFactoryUri = UriUtils.buildUri(
                 this.host.getUri(), ExampleService.FACTORY_LINK);
+        URI exampleODLFactoryUri = UriUtils.buildUri(
+                this.host.getUri(), ExampleODLService.FACTORY_LINK);
         this.host.waitForReplicatedFactoryServiceAvailable(
                 exampleFactoryUri);
+        this.host.waitForReplicatedFactoryServiceAvailable(
+                exampleODLFactoryUri);
     }
 
     private void setUpMultiNode() throws Throwable {
@@ -152,7 +160,7 @@ public class TestSynchronizationTaskService extends BasicTestCase {
                 this.host.getPeerServiceUri(InMemoryExampleService.FACTORY_LINK));
         this.host.waitForReplicatedFactoryServiceAvailable(inMemoryExampleFactoryUri);
         URI ODLExampleFactoryUri = UriUtils.buildUri(
-                this.host.getPeerServiceUri(InMemoryExampleService.FACTORY_LINK));
+                this.host.getPeerServiceUri(ExampleODLService.FACTORY_LINK));
         this.host.waitForReplicatedFactoryServiceAvailable(ODLExampleFactoryUri);
 
     }
@@ -165,18 +173,22 @@ public class TestSynchronizationTaskService extends BasicTestCase {
 
     @Test
     public void ownershipValidation() throws Throwable {
+        System.setProperty(
+                SynchronizationTaskService.PROPERTY_NAME_ENABLE_ODL_SYNCHRONIZATION, "true");
         // This test verifies that only the owner node
         // executes the synchronization task. If the task
         // is started on a non-owner node, the task should
         // self-cancel.
         setUpMultiNode();
+
         ownershipValidationDo(ExampleService.FACTORY_LINK);
         ownershipValidationDo(InMemoryExampleService.FACTORY_LINK);
+        ownershipValidationDo(ExampleODLService.FACTORY_LINK);
     }
 
     public void ownershipValidationDo(String factoryLink) throws Throwable {
-        this.host.createExampleServices(this.host.getPeerHost(), this.serviceCount, null, false, factoryLink);
-
+        boolean skipAvailabilityCheck = factoryLink.equals(ExampleODLService.FACTORY_LINK) ? true : false;
+        this.host.createExampleServices(this.host.getPeerHost(), this.serviceCount, null, skipAvailabilityCheck, factoryLink);
         long membershipUpdateTimeMicros = getLatestMembershipUpdateTime(this.host.getPeerHostUri());
 
         SynchronizationTaskService.State task = createSynchronizationTaskState(membershipUpdateTimeMicros, factoryLink);
@@ -190,7 +202,6 @@ public class TestSynchronizationTaskService extends BasicTestCase {
         TestRequestSender sender = new TestRequestSender(this.host);
         List<SynchronizationTaskService.State> results = sender
                 .sendAndWait(ops, SynchronizationTaskService.State.class);
-
         int finishedCount = 0;
         for (SynchronizationTaskService.State r : results) {
             assertTrue(r.taskInfo.stage == TaskState.TaskStage.FINISHED ||
@@ -290,12 +301,16 @@ public class TestSynchronizationTaskService extends BasicTestCase {
 
     @Test
     public void synchCounts() throws Throwable {
+        System.setProperty(
+                SynchronizationTaskService.PROPERTY_NAME_ENABLE_ODL_SYNCHRONIZATION, "true");
         synchCountsDo(ExampleService.FACTORY_LINK);
         synchCountsDo(InMemoryExampleService.FACTORY_LINK);
+        synchCountsDo(ExampleODLService.FACTORY_LINK);
     }
 
     public void synchCountsDo(String factoryLink) throws Throwable {
-        this.host.createExampleServices(this.host, this.serviceCount, null, false, factoryLink);
+        boolean skipAvailabilityCheck = factoryLink.equals(ExampleODLService.FACTORY_LINK) ? true : false;
+        this.host.createExampleServices(this.host, this.serviceCount, null, skipAvailabilityCheck, factoryLink);
         SynchronizationTaskService.State task = createSynchronizationTaskState(Long.MAX_VALUE, factoryLink);
 
         // Add pagination in query results.
@@ -308,12 +323,14 @@ public class TestSynchronizationTaskService extends BasicTestCase {
         TestRequestSender sender = new TestRequestSender(this.host);
         SynchronizationTaskService.State result = sender
                 .sendAndWait(op, SynchronizationTaskService.State.class);
-
-        assertTrue (result.taskInfo.stage == TaskState.TaskStage.FINISHED);
-        assertTrue (result.synchCompletionCount == this.serviceCount);
+        this.host.waitForTask(SynchronizationTaskService.State.class, result.documentSelfLink, TaskState.TaskStage.FINISHED);
+        op = Operation
+                .createGet(UriUtils.buildUri(this.host, result.documentSelfLink));
+        result = sender.sendAndWait(op, SynchronizationTaskService.State.class);
+        assertTrue(result.synchCompletionCount == this.serviceCount);
 
         // Restart the task to verify counter was reset.
-        task = createSynchronizationTaskState(Long.MAX_VALUE);
+        task = createSynchronizationTaskState(Long.MAX_VALUE, factoryLink);
         task.queryResultLimit = this.serviceCount / 2;
 
         op = Operation
@@ -321,7 +338,10 @@ public class TestSynchronizationTaskService extends BasicTestCase {
                 .setBody(task);
         result = sender.sendAndWait(op, SynchronizationTaskService.State.class);
 
-        assertTrue(result.taskInfo.stage == TaskState.TaskStage.FINISHED);
+        this.host.waitForTask(SynchronizationTaskService.State.class, result.documentSelfLink, TaskState.TaskStage.FINISHED);
+        op = Operation
+                .createGet(UriUtils.buildUri(this.host, result.documentSelfLink));
+        result = sender.sendAndWait(op, SynchronizationTaskService.State.class);
         assertTrue(result.synchCompletionCount == this.serviceCount);
     }
 
@@ -469,13 +489,6 @@ public class TestSynchronizationTaskService extends BasicTestCase {
         consistentStateAfterOwnerStop(InMemoryExampleService.FACTORY_LINK);
     }
 
-    @Ignore("ODL synchronization is yet to be supported")
-    @Test
-    public void consistentStateAfterOwnerStopsODLService() throws Throwable {
-        setUpMultiNode();
-        consistentStateAfterOwnerStop(ExampleODLService.FACTORY_LINK);
-    }
-
     public void consistentStateAfterOwnerStop(
             String factoryLink) throws Throwable {
         long patchCount = 5;
@@ -553,6 +566,104 @@ public class TestSynchronizationTaskService extends BasicTestCase {
         this.host.addPeerNode(hostToRestart);
         this.host.joinNodesAndVerifyConvergence(this.nodeCount);
         return hostToRestart;
+    }
+
+    @Test
+    public void consistentStateAfterOwnerStopsODLServiceWithSync() throws Throwable {
+        System.setProperty(
+                SynchronizationTaskService.PROPERTY_NAME_ENABLE_ODL_SYNCHRONIZATION, "true");
+        setUpMultiNode();
+        consistentOdlStateAfterOwnerStop(ExampleODLService.FACTORY_LINK);
+    }
+
+    @Test
+    public void consistentStateAfterOwnerStopsODLServiceWithoutSync() throws Throwable {
+        System.setProperty(
+                SynchronizationTaskService.PROPERTY_NAME_ENABLE_ODL_SYNCHRONIZATION, "false");
+        setUpMultiNode();
+        consistentOdlStateAfterOwnerStop(ExampleODLService.FACTORY_LINK);
+    }
+
+    public void consistentOdlStateAfterOwnerStop(
+            String factoryLink) throws Throwable {
+        long patchCount = 5;
+        TestRequestSender sender = new TestRequestSender(this.host);
+        this.host.setNodeGroupQuorum(this.nodeCount);
+        this.host.waitForNodeGroupConvergence();
+
+        List<ExampleServiceState> exampleStates = this.host.createExampleServices(
+                this.host.getPeerHost(), this.serviceCount, null, factoryLink);
+
+        VerificationHost owner = this.host.getPeerHost();
+
+        // Send updates to all services and check consistency after owner stops
+        for (ExampleServiceState st : exampleStates) {
+            for (int i = 0; i < patchCount; i++) {
+                URI serviceUri = UriUtils.buildUri(owner, st.documentSelfLink);
+                ExampleServiceState s = new ExampleServiceState();
+                s.counter = (long) i + st.counter;
+                Operation patch = Operation.createPatch(serviceUri).setBody(s);
+                sender.sendAndWait(patch);
+            }
+        }
+
+        // Stop the current owner and make sure that new owner is selected and state is consistent
+        this.host.setNodeGroupQuorum(this.nodeCount - 1);
+        this.host.waitForNodeGroupConvergence(this.nodeCount - 1);
+
+        this.host.stopHost(owner);
+        this.host.waitForNodeGroupConvergence(this.nodeCount - 1);
+        VerificationHost peer = this.host.getPeerHost();
+
+        URI ODLExampleFactoryUri = UriUtils.buildUri(
+                this.host.getPeerServiceUri(ExampleODLService.FACTORY_LINK));
+        this.host.waitForReplicatedFactoryServiceAvailable(ODLExampleFactoryUri);
+
+        //find owner of factory
+        VerificationHost factoryOwner = null;
+        for (VerificationHost host : this.host.getInProcessHostMap().values()) {
+            if (host.isOwner(ExampleODLService.FACTORY_LINK, null)) {
+                // if and only if one owner per document
+                assertNull(factoryOwner);
+                factoryOwner = host;
+            }
+        }
+        assertNotNull(factoryOwner);
+        URI synchronizationTaskUri = UriUtils.buildUri(factoryOwner,
+                ServiceUriPaths.SYNCHRONIZATION_TASKS + "/" + UriUtils.convertPathCharsFromLink(ExampleODLService.FACTORY_LINK));
+        boolean enableODLSynchronization = Boolean.valueOf(System.getProperty(SynchronizationTaskService.PROPERTY_NAME_ENABLE_ODL_SYNCHRONIZATION));
+        TaskState.TaskStage expectedStage = enableODLSynchronization ? TaskState.TaskStage.FINISHED : TaskState.TaskStage.CREATED;
+        this.host.waitFor("Not converged", () -> {
+            Operation getSyncTask = Operation.createGet(synchronizationTaskUri);
+            SynchronizationTaskService.State syncTaskState = sender.sendAndWait(getSyncTask, SynchronizationTaskService.State.class);
+            if (!syncTaskState.taskInfo.stage.equals(expectedStage)) {
+                this.host.log(String.format("actual: %s, expected: %s", syncTaskState.taskInfo.stage.toString(), expectedStage.toString()));
+            }
+            return syncTaskState.taskInfo.stage.equals(expectedStage);
+        });
+
+        if (enableODLSynchronization) {
+            this.host.waitFor("Not converged", () -> {
+                URI factoryUri = UriUtils.buildUri(peer, factoryLink);
+                Operation queryFactory = Operation.createGet(UriUtils.buildExpandLinksQueryUri(factoryUri));
+                ServiceDocumentQueryResult result = sender.sendAndWait(queryFactory, ServiceDocumentQueryResult.class);
+                if (result.documentCount.longValue() != this.serviceCount) {
+                    this.host.log("expected documentCount %d, actual documentCount %d", this.serviceCount, result.documentCount.longValue());
+                    return false;
+                }
+                this.host.log("stopped owner %s, peer %s", owner.getId(), peer.getId());
+                for (Object d : result.documents.values()) {
+                    ExampleServiceState s = Utils.fromJson(d, ExampleServiceState.class);
+                    // synchronization task triggered and owner reselected
+                    if (s.documentOwner.equals(owner.getId())) {
+                        this.host.log("link: %s, owner %s not expected, version %d, epochs %d",
+                                s.documentSelfLink, s.documentOwner, s.documentVersion, s.documentEpoch.longValue());
+                        return false;
+                    }
+                }
+                return true;
+            });
+        }
     }
 
     @Test